api/hotelReservation/requirements.txt
asgiref==3.7.2
attrs==23.1.0
boto3==1.28.28
botocore==1.31.28
cloudpickle==2.2.1
contextlib2==21.6.0
python-decouple==3.8
dill==0.3.7
Django==4.2.4
djangorestframework==3.14.0
google-pasta==0.2.0
importlib-metadata==6.8.0
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
multiprocess==0.70.15
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pathos==0.3.1
platformdirs==3.10.0
pox==0.3.3
ppft==1.7.6.7
protobuf==4.24.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
s3transfer==0.6.2
sagemaker==2.177.1
schema==0.7.5
six==1.16.0
smdebug-rulesconfig==1.0.1
sqlparse==0.4.4
tblib==1.7.0
tzdata==2023.3
urllib3==1.26.16
zipp==3.16.2
